fix(set): set will overwrite value, close #63
such as code #set($a.b.c1 = 1 ) , and with context { a: b: {c2: 2} } .
After set value, context.a.b will be {c1: 1}, c2 is overwrite by #set
