+++ tests/compile.js
@@ -188,121 +188,6 @@ describe('Compile', function() {
     })
   })
 
-  describe('Set && Expression', function() {
-
-    it('set equal to reference', function() {
-      var vm = '#set( $monkey = $bill ) ## variable reference'
-      assert.equal("hello", getContext(vm, {bill: 'hello'}).monkey)
-    })
-
-    it('empty map', function() {
-      var vm = '#set($foo = {})'
-      assert.deepEqual({}, getContext(vm).foo)
-    })
-
-    it('#set array', function() {
-      var vm = '#set($foo = []) #set($foo[0] = 12)'
-      assert.equal(12, getContext(vm).foo[0])
-    })
-
-    it('set equal to literal', function() {
-      var vm = "#set( $monkey.Friend = 'monica' ) ## string literal\n" +
-               '#set( $monkey.Number = 123 ) ##number literal'
-      assert.equal("monica", getContext(vm).monkey.Friend)
-      assert.equal("123", getContext(vm).monkey.Number)
-    })
-    
-    it('set equal to result of method ', function () {
-      var vm = "#set( $monkey = 'monica' ) ## string literal\n" +
-               '#set( $result = $monkey.substring(1) ) ##calling method'
-      assert.equal("monica", getContext(vm).monkey)
-      assert.equal("onica", getContext(vm).result)
-    })
-
-    it('set equal to result of method ', function () {
-      var vm = "#set( $monkey = 1234 ) ## number literal\n" +
-               '#set( $result = $monkey.toString() ) ##calling method'
-      assert.equal("1234", getContext(vm).monkey)
-      assert.equal("1234", getContext(vm).result)
-    })
-
-    it('equal to method/property reference', function() {
-      var vm = "#set($monkey.Blame = $spindoctor.Leak) ## property \n" +
-               '#set( $monkey.Plan = $spindoctor.weave($web) ) ## method'
-      var obj = {
-        spindoctor: {
-          weave: function(name) {
-            return name
-          },
-          Leak: "hello world"
-        },
-        web: "name"
-      }
-
-      assert.equal("hello world", getContext(vm, obj).monkey.Blame)
-      assert.equal("name", getContext(vm, obj).monkey.Plan)
-    })
-
-
-    it('equal to map/list', function() {
-      var vms = [
-        '#set( $monkey.Say = ["Not", $my, "fault"] ) ## ArrayList',
-        '#set( $monkey.Map = {"banana" : "good", "roast beef" : "bad"}) ## Map'
-      ]
-
-      var list = ["Not", "my", "fault"]
-      var map = {banana: "good", 'roast beef': "bad"}
-      assert.deepEqual(list, getContext(vms[0], {my: "my"}).monkey.Say)
-      assert.deepEqual(map, getContext(vms[1]).monkey.Map)
-    })
-
-    it('expression simple math', function() {
-      assert.equal(10, getContext('#set($foo = 2 * 5)').foo)
-      assert.equal(2, getContext('#set($foo = 4 / 2)').foo)
-      assert.equal(-3, getContext('#set($foo = 2 - 5)').foo)
-      assert.equal(1, getContext('#set($foo = 5 % 2)').foo)
-      assert.equal(7, getContext('#set($foo = 7)').foo)
-    })
-
-    it('math with decimal', function() {
-      assert.equal(10.5, getContext('#set($foo = 2.1 * 5)').foo)
-      assert.equal(2.1, getContext('#set($foo = 4.2 / 2)').foo)
-      assert.equal(-7.5, getContext('#set($foo = - 2.5 - 5)').foo)
-    })
-
-    it('expression complex math', function() {
-      assert.equal(20, getContext('#set($foo = (7 + 3) * (10 - 8))').foo)
-      assert.equal(-20, getContext('#set($foo = -(7 + 3) * (10 - 8))').foo)
-      assert.equal(-1, getContext('#set($foo = -7 + 3 * (10 - 8))').foo)
-    })
-
-    it('expression compare', function() {
-      assert.equal(false, getContext('#set($foo = 10 > 11)').foo)
-      assert.equal(true, getContext('#set($foo = 10 < 11)').foo)
-      assert.equal(true, getContext('#set($foo = 10 != 11)').foo)
-      assert.equal(true, getContext('#set($foo = 10 <= 11)').foo)
-      assert.equal(true, getContext('#set($foo = 11 <= 11)').foo)
-      assert.equal(false, getContext('#set($foo = 12 <= 11)').foo)
-      assert.equal(true, getContext('#set($foo = 12 >= 11)').foo)
-      assert.equal(false, getContext('#set($foo = 10 == 11)').foo)
-    })
-
-    it('expression logic', function() {
-      assert.equal(false, getContext('#set($foo = 10 == 11 && 3 > 1)').foo)
-      assert.equal(true, getContext('#set($foo = 10 < 11 && 3 > 1)').foo)
-      assert.equal(true, getContext('#set($foo = 10 > 11 || 3 > 1)').foo)
-      assert.equal(true, getContext('#set($foo = !(10 > 11) && 3 > 1)').foo)
-      assert.equal(false, getContext('#set($foo = $a > $b)', {a: 1, b: 2}).foo)
-      assert.equal(false, getContext('#set($foo = $a && $b)', {a: 1, b: 0}).foo)
-      assert.equal(true, getContext('#set($foo = $a || $b)', {a: 1, b: 0}).foo)
-    })
-
-    it('#set context should be global, #25', function() {
-      var vm = '#macro(local) #set($val =1) $val #end #local() $val'
-      var ret = render(vm).replace(/\s+/g, '')
-      assert.equal('11', ret)
-    })
-  })
 
   describe('Literals', function() {
